+++ backend/command/cmd_install.py
@@ -1,8 +1,15 @@
 import os
 import click
+from pathlib import Path
+
+import subprocess
+
+
 from flask import Flask
 from flask.cli import AppGroup, with_appcontext
 from sqlalchemy import and_
+from sqlalchemy.sql import text
+from sqlalchemy.exc import IntegrityError
 
 from geonature.utils.env import DB
 from pypnnomenclature.models import TNomenclatures, BibNomenclaturesTypes
@@ -10,7 +17,7 @@ from pypnnomenclature.models import TNomenclatures, BibNomenclaturesTypes
 from ..models.monitoring import TMonitoringModules
 from ..config.repositories import config_param, get_config
 from ..config.utils import json_from_file, CONFIG_PATH
-from ..modules.repositories import get_module
+from ..modules.repositories import get_module, get_simple_module, get_source_by_code
 
 app = Flask(__name__)
 monitorings_cli = AppGroup('monitorings')
@@ -24,13 +31,19 @@ def install_monitoring_module(module_config_dir_path, module_path):
     '''
         Module de suivi générique : installation d'un sous module
 
-        Commande d'inst
+        Commande d'installation
+        params :
+            - module_config_dir_path (str) : chemin du répertoire
+                    où se situe les fichiers de configuration du module
+            - module_path (str): code du module
     '''
 
     print('Install module {}'.format(module_path))
 
+    module_monitoring = get_simple_module('module_code', 'MONITORINGS')
+
     try:
-        module = get_module('module_path', module_path)
+        module = get_simple_module('module_path', module_path)
         # test si le module existe
         if(module):
             print("Le module {} existe déjà".format(module_path))
@@ -40,7 +53,11 @@ def install_monitoring_module(module_config_dir_path, module_path):
         pass
 
     if not os.path.exists(module_config_dir_path):
-        print('module_config_dir_path_does not exist (use absolute path)'.format(module_config_dir_path))
+        print(
+            'module_config_dir_path_does not exist (use absolute path)'.format(
+                module_config_dir_path
+            )
+        )
         return
 
     # symlink to config dir
@@ -54,7 +71,11 @@ def install_monitoring_module(module_config_dir_path, module_path):
         )
 
     if not get_config(module_path):
-        print('config directory for module {} does not exist'.format(module_path))
+        print(
+            'config directory for module {} does not exist'.format(
+                module_path
+            )
+        )
         return None
 
     module_desc = config_param(module_path, 'module', 'module_desc')
@@ -69,7 +90,7 @@ et module_desc dans le fichier <dir_module_suivi>/config/monitoring/module.json"
 
     module_data = {
         'module_path': module_path,
-        'module_code': module_path,
+        'module_code': module_path, # ??? Pourquoi module_code = module_path
         'module_label': module_label,
         'module_desc': module_desc,
         'active_frontend': False,
@@ -82,12 +103,121 @@ et module_desc dans le fichier <dir_module_suivi>/config/monitoring/module.json"
     DB.session.add(module)
     DB.session.commit()
 
-    # TODO insert nomenclature
+    #  run specific sql
+    if os.path.exists(module_config_dir_path + '/synthese.sql'):
+        print('Execution du script synthese.sql')
+        sql_script = module_config_dir_path + '/synthese.sql'
+        try:
+            DB.engine.execute(
+                text(
+                    open(sql_script, 'r').read()
+                ).execution_options(autocommit=True)
+            )
+        except Exception as e:
+            print(e)
+            print("Erreur dans le script synthese.sql")
+
+    # insert nomenclature
     add_nomenclature(module_path)
+
+    # creation source pour la synthese
+    txt = ("""
+    INSERT INTO gn_synthese.t_sources(
+        name_source,
+        desc_source,
+        entity_source_pk_field,
+        url_source
+    )
+    VALUES (
+        'MONITORING_{0}',
+        'Données issues du module de suivi générique (sous-module: {1})',
+        'gn_monitoring.vs_{2}.entity_source_pk_value',
+        '#/{3}/object/{2}/visit/observation'
+    );
+        """.format(
+            module_path.upper(), # MONITORING_TEST
+            module_label.lower(), # module de test
+            module.module_path, # test
+            module_monitoring.module_path, # monitorings
+        )
+    )
+
+    DB.engine.execution_options(autocommit=True).execute(txt)
+
+    # exec geonature (update img)
+    subprocess.call(
+        [
+            "geonature update_module_configuration {}"
+            .format(module_monitoring.module_code)
+        ],
+        shell=True)
 
     # TODO ++++ create specific tables
 
     return
+
+
+@monitorings_cli.command('remove')
+@click.argument('module_path')
+@with_appcontext
+def remove_monitoring_module_cmd(module_path):
+    '''
+        Module de suivi générique : suppression d'un sous module
+
+        Commande d'installation
+        params :
+            - module_path (str): code du module
+    '''
+
+    print('Remove module {}'.format(module_path))
+    remove_monitoring_module(module_path)
+
+
+def remove_monitoring_module(module_path):
+    try:
+        module = get_module('module_path', module_path)
+    except Exception:
+        print("le module n'existe pas")
+        return
+
+    # remove module in db
+    try:
+        DB.session.delete(module)
+        DB.session.commit()
+    except IntegrityError as ie:
+        print("Impossible de supprimer le module car il y a des données associées")
+        return
+    except Exception as e:
+        print("Impossible de supprimer le module")
+        raise(e)
+
+    # remove symlink config
+    if os.path.exists(CONFIG_PATH + '/' + module_path):
+        removesymlink(CONFIG_PATH + '/' + module_path)
+
+    # remove symlink image menu modules de suivi
+    img_link = CONFIG_PATH + '/../../frontend/assets/' + module_path + '.jpg'
+    if os.path.exists(img_link):
+        removesymlink(img_link)
+
+    # suppression source pour la synthese
+    source = get_source_by_code("MONITORING_" + module_path.upper())
+    try:
+        DB.session.delete(source)
+        DB.session.commit()
+    except Exception:
+        print("Impossible de supprimer la source")
+        return
+
+    # run specific sql TODO
+    # remove nomenclature TODO
+    return
+
+
+def removesymlink(path):
+    if(os.path.islink(path)):
+        print('remove link ' + path)
+        os.remove(path)
 
 
 def symlink(path_source, path_dest):