+++ frontend/app/services/monitoring-object.service.ts
@@ -1,12 +1,12 @@
-import { MonitoringObject } from './../class/monitoring-object';
-import { Injectable } from '@angular/core';
-import { Observable, of } from '@librairies/rxjs';
+import { MonitoringObject } from "./../class/monitoring-object";
+import { Injectable } from "@angular/core";
+import { Observable, of } from "@librairies/rxjs";
 
-import { ConfigService } from './config.service';
-import { DataMonitoringObjectService } from './data-monitoring-object.service';
-import { DataUtilsService } from './data-utils.service';
-import { Utils } from '../utils/utils';
-import { mergeMap } from '@librairies/rxjs/operators';
+import { ConfigService } from "./config.service";
+import { DataMonitoringObjectService } from "./data-monitoring-object.service";
+import { DataUtilsService } from "./data-utils.service";
+import { Utils } from "../utils/utils";
+import { mergeMap } from "@librairies/rxjs/operators";
 
 @Injectable()
 export class MonitoringObjectService {
@@ -20,13 +20,13 @@ export class MonitoringObjectService {
 
   configUtilsDict = {
     user: {
-      fieldName: 'nom_complet',
+      fieldName: "nom_complet",
     },
     nomenclature: {
-      fieldName: 'label_fr',
+      fieldName: "label_fr",
     },
     taxonomy: {
-      fieldName: 'nom_vern',
+      fieldName: "nom_vern,lb_nom",
     },
   };
 
@@ -38,7 +38,7 @@ export class MonitoringObjectService {
     }
     cache = cache[objectType] = cache[objectType] || {};
 
-    if (objectType === 'module') {
+    if (objectType === "module") {
       return cache;
     }
 
@@ -51,17 +51,17 @@ export class MonitoringObjectService {
   setCache(obj: MonitoringObject, objData) {
     // post ou update
 
-    if (obj.objectType === 'module' && !obj.modulePath) {
+    if (obj.objectType === "module" && !obj.modulePath) {
       return;
     }
-    if (obj.objectType !== 'module' && !obj.id) {
+    if (obj.objectType !== "module" && !obj.id) {
       return;
     }
 
     // object
-    if (obj.objectType === 'module') {
+    if (obj.objectType === "module") {
       const cache = this.cache(obj.modulePath);
-      cache['module'] = objData;
+      cache["module"] = objData;
     } else {
       const cache = this.cache(obj.modulePath, obj.objectType);
       cache[obj.id] = objData;
@@ -79,6 +79,13 @@ export class MonitoringObjectService {
     }
 
     // parent
+    obj.getParent(1).subscribe(() => {
+
+      this.setParentCache(obj, objData);
+    });
+  }
+
+  setParentCache(obj: MonitoringObject, objData) {
     const parent = this.getParentFromCache(obj);
     if (parent) {
       const index = parent.children[obj.objectType].findIndex((child) => {
@@ -93,12 +100,12 @@ export class MonitoringObjectService {
 
         // update nb_child
         const key = Object.keys(parent.properties).find((key) =>
-          ['nb_visits', 'nb_observations', 'nb_sites'].includes(key)
+          ["nb_visits", "nb_observations", "nb_sites"].includes(key)
         );
         if (key) {
+          console.log("up cache", parent.properties.base_site_name, key);
           parent.properties[key] = parent.children[obj.objectType].length;
         }
-        console.log('cache', parent);
       }
     }
   }
@@ -113,6 +120,8 @@ export class MonitoringObjectService {
       obj.parentId
     );
     if (!(parentData && parentData.children)) {
+      console.log(parentData);
+      console.log(parentData && parentData.children);
       return;
     }
     return parentData ? parentData : null;
@@ -120,10 +129,10 @@ export class MonitoringObjectService {
 
   getFromCache(obj: MonitoringObject) {
     // get
-    if (obj.objectType === 'module' && !obj.modulePath) {
+    if (obj.objectType === "module" && !obj.modulePath) {
       return;
     }
-    if (obj.objectType !== 'module' && !obj.id) {
+    if (obj.objectType !== "module" && !obj.id) {
       return;
     }
 
@@ -171,7 +180,7 @@ export class MonitoringObjectService {
       parent.children[obj.objectType].splice(index, 1);
       // update nb_child
       const key = Object.keys(parent.properties).find((key) =>
-        ['nb_visits', 'nb_observations', 'nb_sites'].includes(key)
+        ["nb_visits", "nb_observations", "nb_sites"].includes(key)
       );
       if (key) {
         parent.properties[key] = parent.children[obj.objectType].length;
@@ -197,7 +206,7 @@ export class MonitoringObjectService {
     x = x === undefined ? null : x;
 
     switch (elem.type_widget) {
-      case 'date': {
+      case "date": {
         const date = new Date(x);
         x = x
           ? {
@@ -208,22 +217,22 @@ export class MonitoringObjectService {
           : null;
         break;
       }
-      case 'observers': {
+      case "observers": {
         x = !(x instanceof Array) ? [x] : x;
         break;
       }
-      case 'taxonomy': {
-        x = x ? this._dataUtilsService.getUtil('taxonomy', x, 'all') : null;
+      case "taxonomy": {
+        x = x ? this._dataUtilsService.getUtil("taxonomy", x, "all") : null;
         break;
       }
     }
 
-    if (elem.type_util === 'nomenclature' && Utils.isObject(x)) {
+    if (elem.type_util === "nomenclature" && Utils.isObject(x)) {
       x = this._dataUtilsService
         .getNomenclature(x.code_nomenclature_type, x.cd_nomenclature)
         .pipe(
           mergeMap((nomenclature) => {
-            return of(nomenclature['id_nomenclature']);
+            return of(nomenclature["id_nomenclature"]);
           })
         );
     }
@@ -237,21 +246,21 @@ export class MonitoringObjectService {
   fromForm(elem, val) {
     let x = val;
     switch (elem.type_widget) {
-      case 'date': {
+      case "date": {
         x =
           x && x.year && x.month && x.day
             ? `${x.year}-${x.month}-${x.day}`
             : null;
         break;
       }
-      case 'observers': {
+      case "observers": {
         x =
           elem.max_length === 1 && x instanceof Array && x.length === 1
             ? x[0]
             : x;
         break;
       }
-      case 'taxonomy': {
+      case "taxonomy": {
         x = x instanceof Object ? x.cd_nom : x;
         break;
       }
@@ -260,7 +269,7 @@ export class MonitoringObjectService {
   }
 
   dateFromString(s_date) {
-    const v_date = s_date.split('/');
+    const v_date = s_date.split("/");
     if (v_date.length !== 3) {
       return null;
     }
@@ -269,9 +278,9 @@ export class MonitoringObjectService {
   }
 
   numberFromString(s) {
-    const v = s.split(' ');
+    const v = s.split(" ");
     const s_n = v[0];
-    const v_n = s_n.split('.');
+    const v_n = s_n.split(".");
     v_n[0] = Number(v_n[0]);
     v_n[1] = Number(v_n[1]);
     return v_n.length > 1 && v_n[0] ? v_n : null;