+++ install_gn_module.py
@@ -15,5 +15,12 @@ def gnmodule_install_app(gn_db, gn_app):
         # Install frontend
         gn_db.session.execute(open(str(ROOT_DIR / "data/schema_suivis_generique.sql"), "r").read())
         gn_db.session.commit()
+        gn_db.session.execute(open(str(ROOT_DIR / "data/vues.sql"), "r").read())
+        gn_db.session.commit()
+        gn_db.session.execute(open(str(ROOT_DIR / "data/delete_synthese.sql"), "r").read())
+        gn_db.session.commit()
+
 
         subprocess.call(["npm install"], cwd=str(ROOT_DIR / "frontend"), shell=True)
+
+